requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
pypdf
faiss-cpu
langchainhub
langsmith
langgraph
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4
altair<5
streamlit
streamlit-chat
langchain_pinecone
firecrawl-py